Fixed conflicts with moving wait_pos and wait_alt into separate functions
