+++ flight/utils/drone_takeoff.py
@@ -3,92 +3,7 @@ import logging
 import math
 import mavsdk as sdk
 from mavsdk import System
-from flight.utils.latlon.lat_lon import LatLon
-from flight import config
-
-up_speed: int = -2
-
-async def wait_alt(drone: System) -> bool:
-    """Checks to see if the drone is near the target altitude"""
-    async for position in drone.telemetry.position():
-        altitude: float = round(position.relative_altitude_m, 2)
-        if altitude >= 2:
-            return True
-
-async def getPosition(drone: System, pylon):
-    """Gets x and y distance between 2 sets of latitude and longitude points"""
-
-    async for gps in drone.telemetry.position():
-        altitude = round(gps.relative_altitude_m, 2)
-
-        # Find the current latitude and longitude positions
-        curr_lat = round(gps.latitude_deg, 8)
-        curr_lon = round(gps.longitude_deg, 8)
-
-        current = LatLon(curr_lat, curr_lon)
-
-        deg_to_pylon = current.heading_initial(pylon)
-        offset_point = pylon.offset(deg_to_pylon + config.DEG_OFFSET, config.OFFSET)
-
-        dist = current.distance(offset_point)
-        deg = current.heading_initial(offset_point)
-
-        x = dist*math.sin(math.radians(deg))*1000
-        y = dist*math.cos(math.radians(deg))*1000
-
-        # x = ((target_lon - curr_lon)* 40000 * math.cos((target_lat + curr_lat) * math.pi / 360) / 360) * 1000
-
-        # y = ((target_lat - curr_lat) * 40000 / 360) * 1000
-
-        return (x, y, altitude)
-
-async def getVelocity(drone: System, pylon):
-    """Returns the needed x-velocity dx, y-velocity \
-       dy, altitude speed, and needed degree deg of \
-       the drone to go to a given position."""
-    """Returns a list of the format (dx, dy, alt, deg)"""
-
-    # Get x and y distance from target lat and lon points, and
-    # altitude
-    position = await getPosition(drone, pylon)
-
-    x = position[0]
-    y = position[1]
-    altitude = position[2]
-
-    if altitude >= 3:
-        alt = 0.2
-    elif altitude <= 2:
-        alt = -0.2
-    else:
-        alt = 0
-
-    # Calculate the degree to point the drone at
-    try:
-        deg = round((((math.atan(x / y) / math.pi) * 180)))
-
-        if y < 0:
-            z = 180
-            z = math.copysign(z, deg)
-            deg = z + deg
-    except ZeroDivisionError:
-        deg = round((((math.asin(x / (math.sqrt((x ** 2) + (y ** 2)))) / math.pi)* 180)))
-
-        if y < 0:
-            z = 180
-            z = math.copysign(z, deg)
-            deg = z + deg
-
-    # Calculate the needed velocity needed to reach the target lat lon points
-    try:
-        dx = math.copysign(35 * math.cos(math.atan(y / x)), x)
-        dy = math.copysign(35 * math.sin(math.atan(y / x)), y)
-
-    except ZeroDivisionError:
-        dx = math.copysign(35 * math.cos(math.asin(y / (math.sqrt((x ** 2) + (y ** 2))))), x)
-        dy = math.copysign(35 * math.sin(math.asin(y / (math.sqrt((x ** 2) + (y ** 2))))), y)
-
-    return (dx, dy, alt, deg)
+from flight.utils.movement import getVelocity, wait_alt
 
 async def takeoff(drone: System, pylon ) -> None:
     """This function expects that the drone is already armed."""
@@ -125,6 +40,8 @@ async def takeoff(drone: System, pylon ) -> None:
     dx = velocity[0]
     # Y-velocity
     dy = velocity[1]
+    # Up speed
+    alt = velocity[2]
     # Degree
     deg = velocity[3]
 
@@ -132,7 +49,7 @@ async def takeoff(drone: System, pylon ) -> None:
 
     # Start the drone pointing in the direction of the first pylon
     await drone.offboard.set_velocity_ned(
-        sdk.VelocityNedYaw(dy, dx, up_speed, deg)
+        sdk.VelocityNedYaw(dy, dx, -.5, deg)
     )
 
     # Loops until the desired altitude has been attained,
